Update i3/mkenv.sh
1. Add check for I3_COLOR_THEME before generate .i3/config
2. Add check for file before copy i3.desktop


Former-commit-id: 1c6bb52bff52650a8b4595ede6411042449a0c90

+++ i3/mkenv.sh
@@ -13,6 +13,9 @@ ln -sfnv $CONFIG_HOME/dunst/dunstrc $DUNST_HOME/dunstrc
 
 mkdir -p $HOME/bin
 ln -sfnv $CONFIG_HOME/i3bang/i3bang.rb $HOME/bin/i3bang
+if [ -z "$I3_COLOR_THEME" ]; then
+  export I3_COLOR_THEME=dark
+fi
 i3bang
 
 # check if 'consolekit' is installed or not
@@ -28,4 +31,6 @@ if [ -z "$PACKAGE" ]; then
   echo 'Install consolekit ... done'
 fi
 
-sudo cp $CONFIG_HOME/xsessions/i3.desktop /usr/share/xsessions/i3.desktop
+if [ -e /usr/share/xsessions/i3.desktop ]; then
+  sudo cp $CONFIG_HOME/xsessions/i3.desktop /usr/share/xsessions/i3.desktop
+fi
